Add instructions for adapting UI components
Add information for how to adapt existing UI components. I am hoping to clarify the extensibility of the existing UI in addition to the creating of new components, since it wasn't immediately clear to me but a few pointers from the Evidence team got me pointed in the right direction.