feat: allow overriding usage of npx
I don't love this solution, but throwing it out there for discussion.

*Problem*
When `workspaces` are in play, `npx` will execute the command with
the current working directory set to the `workspace` directory.

```
/root
  package.json
 packages
   my-evidence-proj <- CWD of the vite command ran by npx (incorrect)
     package.json
     pages
     evidence.plugins.yaml
     ...
     .evidence
       template <- CWD of the npx command (correct)
       tailwind.config.cjs
```

- https://github.com/npm/cli/blob/latest/lib/npm.js#L233-L248
- https://github.com/npm/cli/blob/latest/lib/commands/exec.js#L34-L38
- https://github.com/npm/cli/blob/latest/lib/commands/exec.js#L103

Which ultimately means that `vite` gets run in the wrong `cwd` and thus
can't find the template / config.

I think arguably this is surprising behavior from `npm` / `npx`, but
changing it would probably be a hard sell. Perhaps adding a flag to
specify the `cwd` might be accepted?

Aside from this, personally I find the use of `npx` for this usecase to
be a bit surprising - it's not clear to me how the dependency versions
are being controlled.

*Possible Solutions*
- Change `npm` to allow specifying the `cwd` when running `npx` / `npm exec`
  in a project using workspaces
- Install all dependencies in project and run using `yarn vite` instead
  of `npx`
- Explicitly pass the template root directory to `vite`
  - this almost works, but `svelte-kit` overrides it / ignores it and
    ends up being unable to find anything
  - if it's possible to explicitly pass the template root to svelte-kit
    instead of relying on the `process.cwd()` that would probably be the
    cleanest way to solve this